Added new constants + Added init files + Updated imports + Minor refactoring + Updated .gitignore
