+++ .gitignore
@@ -1,9 +1,11 @@
-.venv
+*.venv
+*.env
 *__pycache__*/
 *.pyc
-.DS_Store
+*.DS_Store
+
 out/checkpoints/*
 out/res/*
-*.env
+
 data/HR/*
 data/LR/*
