+++ src/pipelines/config.py
@@ -1,14 +1,12 @@
 from sentinelhub import SHConfig
-from oauthlib.oauth2 import BackendApplicationClient
-from requests_oauthlib import OAuth2Session
 from dotenv import load_dotenv
 
 import os
 import logging
 
 # Setup logging
-logging.basicConfig(level=logging.DEBUG)
-logging.captureWarnings(True)
+# logging.basicConfig(level=logging.DEBUG)
+# logging.captureWarnings(True)
 
 load_dotenv()
 
@@ -31,25 +29,3 @@ config.sh_token_url = 'https://identity.dataspace.copernicus.eu/auth/realms/CDSE
 config.save(CONFIG_NAME)
 
 print(SHConfig())
-
-# def sentinelhub_compliance_hook(response):
-#     '''
-#     Check response status and raise an error if needed.
-#     '''
-#     response.raise_for_status()
-#     return response
-
-# # Create a session
-# client = BackendApplicationClient(client_id=client_id)
-# oauth = OAuth2Session(client=client)
-
-# # Get token for the session
-# token = oauth.fetch_token(token_url='https://identity.dataspace.copernicus.eu/auth/realms/CDSE/protocol/openid-connect/token',
-#                           client_secret=client_secret, include_client_id=True)
-
-# # Register the compliance hook to check responses before sending back to user
-# oauth.register_compliance_hook("access_token_response", sentinelhub_compliance_hook)
-
-# # All requests using this session will have an access token automatically added
-# resp = oauth.get("https://sh.dataspace.copernicus.eu/configuration/v1/wms/instances")
-# print(resp.content)
