+++ src/utils/utils.py
@@ -1,6 +1,5 @@
 import math
 from pathlib import Path
-from typing import Tuple
 
 import cv2
 import numpy as np
